+++ Documentation/Engineering/Swift64/Probes/Ownership/F04_unique_box.swift
@@ -1,0 +1,10 @@
+
+struct Element: ~Copyable { var number: Int }
+func run() {
+    var box = UniqueBox(Element(number: 3))
+    box.value.number += 1
+    let value = box.consume()
+    precondition(value.number == 4)
+    print("F04: unique box mutation and consume passed")
+}
+run()