+++ Documentation/Engineering/Swift64/Probes/Ownership/F10_optional_put.swift
@@ -1,0 +1,6 @@
+
+struct Element: ~Copyable { var number: Int }
+func trial(_ optional: inout Element?) {
+    var reference = optional.put(Element(number: 3))
+    reference.value.number += 1
+}